target+source: move `IntTy` and `FloatTy` into `hash-target`
