+++ compiler/hash-source/src/constant.rs
@@ -4,36 +4,18 @@
 //! be passed around each stage of the compiler and can just
 //! be interned, and accessed when needed.
 
-use std::{
-    cmp::Ordering,
-    fmt::{self, Display},
-    ops::Neg,
-};
+use std::{cmp::Ordering, fmt, ops::Neg};
 
 use dashmap::DashMap;
 use fnv::FnvBuildHasher;
-use hash_target::size::Size;
+// Re-export the "primitives" from the hash-target crate so that everyone can use
+// them who depends on `hash-source`
+pub use hash_target::primitives::{FloatTy, IntTy, SIntTy, UIntTy};
 use hash_utils::counter;
 use lazy_static::lazy_static;
 use num_bigint::BigInt;
 
 use crate::identifier::{Identifier, IDENTS};
-
-/// The type of the float the [FloatLit] is storing.
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-pub enum FloatTy {
-    F32,
-    F64,
-}
-
-impl Display for FloatTy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            FloatTy::F32 => write!(f, "f32"),
-            FloatTy::F64 => write!(f, "f64"),
-        }
-    }
-}
 
 /// The inner stored value of a [FloatConstant].
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd)]
@@ -42,8 +24,8 @@ pub enum FloatConstantValue {
     F32(f32),
 }
 
-impl Display for FloatConstantValue {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for FloatConstantValue {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             FloatConstantValue::F64(inner) => write!(f, "{inner}"),
             FloatConstantValue::F32(inner) => write!(f, "{inner}"),
@@ -106,8 +88,8 @@ counter! {
     method_visibility:,
 }
 
-impl Display for FloatConstant {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for FloatConstant {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.value)?;
 
         if let Some(suffix) = self.suffix {
@@ -118,99 +100,11 @@ impl Display for FloatConstant {
     }
 }
 
-impl Display for InternedFloat {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for InternedFloat {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", CONSTANT_MAP.lookup_float_constant(*self))
     }
 }
-
-/// Unsigned integer type variants.
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
-pub enum UIntTy {
-    U8,
-    U16,
-    U32,
-    U64,
-    U128,
-    USize,
-    UBig,
-}
-
-impl UIntTy {
-    /// Get the size of [IntTy] in bytes. Returns [None] for
-    /// [UIntTy::UBig] variants
-    pub fn size(&self, ptr_width: usize) -> Option<Size> {
-        match self {
-            UIntTy::U8 => Some(Size::from_bytes(1)),
-            UIntTy::U16 => Some(Size::from_bytes(2)),
-            UIntTy::U32 => Some(Size::from_bytes(4)),
-            UIntTy::U64 => Some(Size::from_bytes(8)),
-            UIntTy::USize => Some(Size::from_bytes(ptr_width)),
-            UIntTy::U128 => Some(Size::from_bytes(16)),
-            UIntTy::UBig => None,
-        }
-    }
-
-    /// Create a new [UnitTy] from a given [Size]. This assumes that
-    /// the maximum passed [Size] can be represented as a [UIntTy::U128].
-    ///
-    /// Additionally, this will never use the `usize` type to avoid confusion
-    /// between different platforms/targets.
-    pub fn from_size(size: Size) -> Self {
-        match size.bytes() {
-            0..=1 => UIntTy::U8,
-            2 => UIntTy::U16,
-            3..=4 => UIntTy::U32,
-            5..=8 => UIntTy::U64,
-            9..=16 => UIntTy::U128,
-            _ => unreachable!(),
-        }
-    }
-
-    /// Function to get the largest possible integer represented within this
-    /// type. For sizes `ibig` and `ubig` there is no defined max and so the
-    /// function returns [None].
-    pub fn max(&self, ptr_width: usize) -> Option<BigInt> {
-        match self {
-            UIntTy::U8 => Some(BigInt::from(u8::MAX)),
-            UIntTy::U16 => Some(BigInt::from(u16::MAX)),
-            UIntTy::U32 => Some(BigInt::from(u32::MAX)),
-            UIntTy::U64 => Some(BigInt::from(u64::MAX)),
-            UIntTy::U128 => Some(BigInt::from(u128::MAX)),
-            UIntTy::USize => {
-                let max = !0u64 >> (64 - (ptr_width * 8));
-                Some(BigInt::from(max))
-            }
-            UIntTy::UBig => None,
-        }
-    }
-
-    /// Function to get the most minimum integer represented within this
-    /// type.
-    pub fn min(&self) -> BigInt {
-        0.into()
-    }
-
-    /// Convert the [UIntTy] into a primitive type name
-    pub fn to_name(&self) -> &'static str {
-        match self {
-            UIntTy::U8 => "u8",
-            UIntTy::U16 => "u16",
-            UIntTy::U32 => "u32",
-            UIntTy::U64 => "u64",
-            UIntTy::U128 => "u128",
-            UIntTy::USize => "usize",
-            UIntTy::UBig => "ubig",
-        }
-    }
-}
-
-impl From<UIntTy> for IntTy {
-    fn from(value: UIntTy) -> Self {
-        IntTy::UInt(value)
-    }
-}
-
 impl From<UIntTy> for Identifier {
     fn from(value: UIntTy) -> Self {
         match value {
@@ -221,90 +115,6 @@ impl From<UIntTy> for Identifier {
             UIntTy::U128 => IDENTS.u128,
             UIntTy::USize => IDENTS.usize,
             UIntTy::UBig => IDENTS.ubig,
-        }
-    }
-}
-
-impl fmt::Display for UIntTy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.to_name())
-    }
-}
-
-/// Signed integer type variants.
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
-pub enum SIntTy {
-    I8,
-    I16,
-    I32,
-    I64,
-    I128,
-    ISize,
-    IBig,
-}
-
-impl SIntTy {
-    /// Get the size of [IntTy] in bytes. Returns [None] for
-    /// [UIntTy::UBig] variants
-    pub fn size(&self, ptr_width: usize) -> Option<Size> {
-        match self {
-            SIntTy::I8 => Some(Size::from_bytes(1)),
-            SIntTy::I16 => Some(Size::from_bytes(2)),
-            SIntTy::I32 => Some(Size::from_bytes(4)),
-            SIntTy::I64 => Some(Size::from_bytes(8)),
-            SIntTy::ISize => Some(Size::from_bytes(ptr_width)),
-            SIntTy::I128 => Some(Size::from_bytes(16)),
-            SIntTy::IBig => None,
-        }
-    }
-
-    /// Function to get the largest possible integer represented within this
-    /// type. For sizes `ibig` and `ubig` there is no defined max and so the
-    /// function returns [None].
-    pub fn max(&self, ptr_width: usize) -> Option<BigInt> {
-        match self {
-            SIntTy::I8 => Some(BigInt::from(i8::MAX)),
-            SIntTy::I16 => Some(BigInt::from(i16::MAX)),
-            SIntTy::I32 => Some(BigInt::from(i32::MAX)),
-            SIntTy::I64 => Some(BigInt::from(i64::MAX)),
-            SIntTy::I128 => Some(BigInt::from(i128::MAX)),
-            SIntTy::ISize => {
-                // convert the size to a signed integer
-                let max = (1u64 << (ptr_width * 8 - 1)) - 1;
-                Some(BigInt::from(max))
-            }
-            SIntTy::IBig => None,
-        }
-    }
-
-    /// Function to get the most minimum integer represented within this
-    /// type. For sizes `ibig` and `ubig` there is no defined minimum and so the
-    /// function returns [None].
-    pub fn min(&self, ptr_width: usize) -> Option<BigInt> {
-        match self {
-            SIntTy::I8 => Some(BigInt::from(i8::MIN)),
-            SIntTy::I16 => Some(BigInt::from(i16::MIN)),
-            SIntTy::I32 => Some(BigInt::from(i32::MIN)),
-            SIntTy::I64 => Some(BigInt::from(i64::MIN)),
-            SIntTy::I128 => Some(BigInt::from(i128::MIN)),
-            SIntTy::ISize => {
-                let min = (i64::MAX) << ((ptr_width * 8) - 1);
-                Some(BigInt::from(min))
-            }
-            SIntTy::IBig => None,
-        }
-    }
-
-    /// Convert the [IntTy] into a primitive type name
-    pub fn to_name(&self) -> &'static str {
-        match self {
-            SIntTy::I8 => "i8",
-            SIntTy::I16 => "i16",
-            SIntTy::I32 => "i32",
-            SIntTy::I64 => "i64",
-            SIntTy::I128 => "i128",
-            SIntTy::ISize => "isize",
-            SIntTy::IBig => "ibig",
         }
     }
 }
@@ -320,80 +130,6 @@ impl From<SIntTy> for Identifier {
             SIntTy::ISize => IDENTS.isize,
             SIntTy::IBig => IDENTS.ibig,
         }
-    }
-}
-
-impl From<SIntTy> for IntTy {
-    fn from(value: SIntTy) -> Self {
-        IntTy::Int(value)
-    }
-}
-
-impl fmt::Display for SIntTy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.to_name())
-    }
-}
-
-/// The representation of an integer type.
-#[derive(Debug, PartialEq, Eq, Clone, Copy, Hash)]
-pub enum IntTy {
-    /// Signed integer variant.
-    Int(SIntTy),
-    /// Unsigned integer variant.
-    UInt(UIntTy),
-}
-
-impl IntTy {
-    /// Convert the type into a name.
-    pub fn to_name(&self) -> &'static str {
-        match self {
-            IntTy::Int(ty) => ty.to_name(),
-            IntTy::UInt(ty) => ty.to_name(),
-        }
-    }
-
-    /// Function to get the largest possible integer represented within this
-    /// type. For sizes `ibig` and `ubig` there is no defined max and so the
-    /// function returns [None].
-    pub fn max(&self, ptr_width: usize) -> Option<BigInt> {
-        match self {
-            IntTy::Int(ty) => ty.max(ptr_width),
-            IntTy::UInt(ty) => ty.max(ptr_width),
-        }
-    }
-
-    /// Function to get the most minimum integer represented within this
-    /// type. For sizes `ibig` there is no defined minimum and so the
-    /// function returns [None].
-    pub fn min(&self, ptr_width: usize) -> Option<BigInt> {
-        match self {
-            IntTy::Int(ty) => ty.min(ptr_width),
-            IntTy::UInt(ty) => Some(ty.min()),
-        }
-    }
-
-    /// Function to get the size of the integer type in bytes.
-    pub fn size(&self, ptr_width: usize) -> Option<Size> {
-        match self {
-            IntTy::Int(ty) => ty.size(ptr_width),
-            IntTy::UInt(ty) => ty.size(ptr_width),
-        }
-    }
-
-    /// Check if the type is signed.
-    pub fn is_signed(&self) -> bool {
-        matches!(self, IntTy::Int(_))
-    }
-
-    /// Check if the type is a pointer integral type, i.e. `isize` or `usize`.
-    pub fn is_ptr_sized_integral(self) -> bool {
-        matches!(self, IntTy::Int(SIntTy::ISize) | IntTy::UInt(UIntTy::USize))
-    }
-
-    /// Check if the type is a [BigInt] variant, i.e. `ibig` or `ubig`.
-    pub fn is_big_sized_integral(self) -> bool {
-        matches!(self, IntTy::Int(SIntTy::IBig) | IntTy::UInt(UIntTy::UBig))
     }
 }
 
@@ -427,12 +163,6 @@ impl TryFrom<Identifier> for IntTy {
             i if i == IDENTS.ubig => Ok(IntTy::UInt(UIntTy::UBig)),
             _ => Err(()),
         }
-    }
-}
-
-impl fmt::Display for IntTy {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.to_name())
     }
 }
 
@@ -702,8 +432,8 @@ counter! {
     method_visibility:,
 }
 
-impl Display for IntConstant {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for IntConstant {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match &self.value {
             // We want to snip the value from the `total` value since we don't care about the
             // rest...
@@ -721,8 +451,8 @@ impl Display for IntConstant {
     }
 }
 
-impl Display for InternedInt {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for InternedInt {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", CONSTANT_MAP.lookup_int_constant(*self))
     }
 }
@@ -734,8 +464,8 @@ counter! {
     method_visibility:,
 }
 
-impl Display for InternedStr {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl fmt::Display for InternedStr {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", CONSTANT_MAP.lookup_string(*self))
     }
 }
@@ -904,11 +634,13 @@ impl ConstantMap {
 
 #[cfg(test)]
 mod tests {
-    use hash_target::size::Size;
+    use hash_target::{
+        primitives::{SIntTy, UIntTy},
+        size::Size,
+    };
     use num_bigint::BigInt;
 
-    use super::{IntConstantValue, SIntTy};
-    use crate::constant::UIntTy;
+    use crate::constant::IntConstantValue;
 
     #[test]
     fn test_max_signed_int_value() {
