+++ compiler/hash-target/Cargo.toml
@@ -3,3 +3,6 @@ name = "hash-target"
 version = "0.1.0"
 authors = ["The Hash Language authors"]
 edition = "2021"
+
+[dependencies]
+num-bigint = "0.4"
