+++ compiler/hash-target/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod abi;
 pub mod alignment;
 pub mod layout;
+pub mod primitives;
 pub mod size;
 
 use std::{
